requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.2
brotli==1.2.0
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.9.0
diff-match-patch==20241021
Django==5.2
django-import-export==4.4.0
fonttools==4.61.1
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.17
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.17
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
